pyproject.toml
[project]
name = "product-intelligence-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=1.9.3",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.129.2",
    "httpx>=0.28.1",
    "openai>=1.83.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "streamlit>=1.54.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.41.0",
]
